requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.14
django-autoslug==1.9.9
django-bootstrap5==22.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
flake8==7.1.0
idna==3.7
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pluggy==1.5.0
pycodestyle==2.12.0
pyflakes==3.2.0
pytest==8.3.1
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
